backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
beautifulsoup4==4.13.4
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
curl_cffi==0.13.0
distro==1.9.0
fastapi==0.116.1
frozendict==2.4.6
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.74
langchain-openai==0.3.31
langchain-text-splitters==0.3.9
langsmith==0.4.15
multitasking==0.0.12
numpy==2.3.2
openai==1.101.0
orjson==3.11.2
packaging==25.0
pandas==2.3.2
peewee==3.18.2
platformdirs==4.3.8
protobuf==6.32.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.5
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.43
starlette==0.47.2
tenacity==9.1.2
tiktoken==0.11.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
yfinance==0.2.65
zstandard==0.24.0
langchain-groq




requests
pandas
numpy
yfinance
TA-Lib-binary==0.4.27 
pandas_ta==0.3.14b0


# LangChain stack
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.42
langgraph==0.2.16
langchain-groq==0.1.5
chromadb==0.5.5
faiss-cpu==1.8.0.post1 # optional if you prefer FAISS over Chroma
fastembed==0.3.3 # high-performance local embeddings


# Observability
langsmith==0.1.106
sentence-transformers==3.3.1
langchain-huggingface==0.1.2